refactor: add todo mapper interface and its impl to decouple the business logic from its implementation with some external lib
